requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
